Refactor project and re-export
