+++ example/lib/main.dart
@@ -1,6 +1,4 @@
-import 'package:flip_card/flip_card_controller.dart';
 import 'package:flutter/material.dart';
-
 import 'package:flip_card/flip_card.dart';
 
 void main() => runApp(const MyApp());
